Move controller-related constants and enums into controllers.h, remove globals.h
